requirements.txt
fastapi==0.115.0
uvicorn==0.35.0
psycopg[binary]==3.3.2
pydantic==2.11.7
pytest==8.3.3
pytest-cov==6.0.0
pytest-playwright==0.7.2
playwright==1.58.0
httpx==0.28.1
flake8==7.1.1
sphinx==9.1.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-instrumentation-fastapi>=0.41b0
opentelemetry-instrumentation-psycopg>=0.41b0
opentelemetry-exporter-otlp-proto-grpc>=1.20.0
opentelemetry-exporter-prometheus>=0.41b0
prometheus-client>=0.20.0
